sf-opportunity-update/requirements.txt
google-cloud-error-reporting
simple-salesforce
google-cloud-secret-manager